+++ templates/sign-up.php
@@ -12,27 +12,27 @@
       <h2>Регистрация нового аккаунта</h2>
       <div class="form__item <?= $errors['email'] ? 'form__item--invalid' : '' ?>"> <!-- form__item--invalid -->
         <label for="email">E-mail*</label>
-        <input id="email" type="text" name="email" placeholder="Введите e-mail" value="<?= $_POST['email'];?>">
+        <input id="email" type="text" name="email" placeholder="Введите e-mail" value="<?= $user_data['email'];?>">
         <span class="form__error"><?= $errors['email'] ?></span>
       </div>
       <div class="form__item <?= $errors['password'] ? 'form__item--invalid' : '' ?>">
         <label for="password">Пароль*</label>
-        <input id="password" type="text" name="password" placeholder="Введите пароль" value="<?= $_POST['password'];?>">
+        <input id="password" type="text" name="password" placeholder="Введите пароль" value="<?= $user_data['password'];?>">
         <span class="form__error"><?= $errors['password'] ?></span>
       </div>
       <div class="form__item <?= $errors['name'] ? 'form__item--invalid' : '' ?>">
         <label for="name">Имя*</label>
-        <input id="name" type="text" name="name" placeholder="Введите имя" value="<?= $_POST['name'];?>">
+        <input id="name" type="text" name="name" placeholder="Введите имя" value="<?= $user_data['name'];?>">
         <span class="form__error"><?= $errors['name'] ?></span>
       </div>
       <div class="form__item <?= $errors['contacts'] ? 'form__item--invalid' : '' ?>">
         <label for="message">Контактные данные*</label>
-        <textarea id="message" name="contacts" placeholder="Напишите как с вами связаться"><?= $_POST['contacts'];?></textarea>
+        <textarea id="message" name="contacts" placeholder="Напишите как с вами связаться"><?= $user_data['contacts'];?></textarea>
         <span class="form__error"><?= $errors['contacts'] ?></span>
       </div>
 
 
-      <div class="form__item form__item--file form__item--last <?= $user_avatar ? 'form__item--invalid' : '' ?>">
+      <div class="form__item form__item--file form__item--last <?= $errors['avatar'] ? 'form__item--invalid' : '' ?>">
         <label>Аватар</label>
         <div class="preview">
           <button class="preview__remove" type="button">x</button>
@@ -45,7 +45,7 @@
           <label for="photo2">
             <span>+ Добавить</span>
           </label>
-        </div><span class="form__error"><?= $user_avatar ?></span>
+        </div><span class="form__error"><?= $errors['avatar'] ?></span>
       </div>
